Mais ajustes

+++ RttBMS/src/objetos/ObjCelula.cpp
@@ -23,11 +23,11 @@ float ObjCelula::getLeituraTensaoAnterior(){
   return _tensao_anterior;
 }
 
-void ObjCelula::setPortaInput(uint8_t porta){
+void ObjCelula::setPortaInput(int porta){
   _porta_input = porta;
 }
 
-uint8_t ObjCelula::getPortaInput(){
+int ObjCelula::getPortaInput(){
   return _porta_input;
 }
