+++ RttBMS/src/objetos/ObjCelula.h
@@ -10,15 +10,15 @@ class ObjCelula {
     void setLeituraTensao(float leitura);
     int getNumeroCelula();
     void setNumeroCelula(int numero_celula);
-    void setPortaInput(uint8_t porta);
-    uint8_t getPortaInput();
+    void setPortaInput(int porta);
+    int getPortaInput();
     void setPortaControle(int porta);
     int getPortaControle();
 
   private:
     //Numero da celula
     int _numero_celula;
-    uint8_t _porta_input;
+    int _porta_input;
     int _porta_controle;
     float _leitura_tensao;
     float _tensao_anterior;
